used util function for time formatting
