+++ client/src/components/posts/AllPosts.scss
@@ -115,7 +115,7 @@ a{
         padding-left: 0.1rem;
         text-align: start;
     }
-    p{
+    .article_content p{
         padding-left: 7px;
         width: 50rem;
         max-width: 53ch;
@@ -123,7 +123,7 @@ a{
 }
 
 @media screen and (min-width:1200px){
-    p{
+    .article_content p{
         padding-left: 7px;
         width: 50rem;
         max-width: 100ch;
